% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasttable.R
\name{contrasttable}
\alias{contrasttable}
\title{Creating counterfactual mobility table based on a comparison case}
\usage{
contrasttable(data, o, d, contrast)
}
\arguments{
\item{data}{Data frame containing the origin and destination variables}

\item{o}{Origin variable name, a string}

\item{d}{Destination variable name, a string}

\item{contrast}{A comparison mobility table with the same origin and destination dimensions}
}
\value{
A counterfactual mobility table with odds ratios from the comparison case and marginal O and D distributions from data
}
\description{
Creating counterfactual mobility table based on a comparison case
}
\examples{
toy_data <- data.frame(
  padeg = factor(sample(1:3, 100, replace = TRUE)),
  degree = factor(sample(1:3, 100, replace = TRUE))
)

comparison_data <- data.frame(
  padeg = factor(sample(1:3, 100, replace = TRUE)),
  degree = factor(sample(1:3, 100, replace = TRUE))
)


comparison_table <- table(comparison_data$padeg,comparison_data$degree)

contrasttable(toy_data, o = "padeg", d = "degree", contrast = comparison_table)

}
