% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepstable.R
\name{stepstable}
\alias{stepstable}
\title{Creating counterfactual mobility table based on specified interventional strength}
\usage{
stepstable(data, o, d, strength)
}
\arguments{
\item{data}{Data frame containing the origin and destination variables}

\item{o}{Origin variable name, a string}

\item{d}{Destination variable name, a string}

\item{strength}{Interventional strength as defined by proportional logged odds-ratio distance from the independence table, range from 0-1}
}
\value{
A counterfactual mobility table as implied by the interventional strength
}
\description{
Creating counterfactual mobility table based on specified interventional strength
}
\examples{
toy_data <- data.frame(
  padeg = factor(sample(1:3, 100, replace = TRUE)),
  degree = factor(sample(1:3, 100, replace = TRUE))
)
stepstable(toy_data, o = "padeg", d = "degree", strength = 1)

}
