% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.R
\docType{data}
\name{toy}
\alias{toy}
\title{Random subset of GSS DATA}
\format{
\subsection{\code{toy}}{

A data frame with 500 rows and 15 columns:
\describe{
\item{age}
\item{sibs}
\item{ethnic}
\item{res16}
\item{reg16}
\item{family16}
\item{relig16}
\item{fund16}
\item{incom16}
\item{pawrkslf}
\item{parborn}
\item{padeg}
\item{madeg}
\item{degree}
\item{childs}
}
}
}
\source{
\if{html}{\out{
<GSS>
}}
}
\usage{
toy
}
\description{
A subset of data GSS that includes origin, destination, fertility, and controls
Report ...
}
\keyword{datasets}
